fixed basic user routes to use database instead of hard coded data
