+++ src/backend/app.js
@@ -1,16 +1,8 @@
 import express from 'express'
 import mongoose from 'mongoose'
 import cors from 'cors'
-import dotenv from 'dotenv'
+import { User } from './db.js'
 
-const users = [
-    {
-        firstName: "John",
-        lastName: "Doe",
-        email: "John@example.com.au",
-        password: "password123",
-        isAdmin: true
-    }]
 
 const hireOptions = [
     {
@@ -23,24 +15,6 @@ const hireOptions = [
     }]
 
 
-dotenv.config()
-
-try {
-    const m = await mongoose.connect(process.env.DB_URI)
-    console.log(m.connection.readyState == 1 ? 'Mongoose Connected' : 'Mongoose failed to connect')
-}
-catch (err) {
-    console.error(err)
-}
-
-const User  = new mongoose.model('User', {
-    firstName: { type: String, required: true },
-    lastName: { type: String, required: true },
-    email: { type: String, required: true, unique: true },
-    password: { type: String, required: true },
-    isAdmin: { type: Boolean, required: true, default: false}
-})
-
 const app = express()
 
 // Middleware
@@ -49,12 +23,17 @@ app.use(express.json())
 
 app.get('/', (req, res) => res.send({ info: 'GC Activity Rentals' }))
 
-app.get('/users', (req, res) => res.send(users))
+// Get Users
+app.get('/users', async (req, res) => res.send(await User.find()))
 
 // Create a new User
 app.post('/users', async (req, res) => {
-    const newUser = await User.create(req.body)
-    res.status(201).send(newUser)
+    try {
+        const newUser = await User.create(req.body)
+        res.status(201).send(newUser)
+    } catch (err) {
+        res.status(400).send({error: err.message})
+    }
 })
 
 app.get('/hireOptions', (req, res) => res.send(hireOptions))
