+++ src/backend/models/users.js
@@ -11,4 +11,4 @@
 
 // const User = mongoose.model("User", userSchema)
 
-// export default User
+// export {User}
